feat(core): add method to dispose models (#6)
* feat(core): expose MonacoEditorService

* feat(MonacoEditorService): add method to dispose models

This method enables consumers to dispose all registered editor models from the
outside.

There are cases where the editor might receive a file with the same name/uri
but different contents and should therefore create a new model for it. Using this
new method, this can be controlled by disposing all models before.